pyproject.toml
[project]
name = "triage-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.39.3",
    "litellm>=1.74.0",
    "python-dotenv>=1.1.1",
    "fastapi[standard]>=0.104.1",
    "uvicorn>=0.24.0",
    "httpx>=0.25.0",
    "loguru>=0.7.3",
]

[dependency-groups]
dev = [
    "killport>=1.2.0",
]
